requirements.txt
Flask==2.3.3
gunicorn==21.2.0
pandas==1.5.3
scikit-learn==1.3.2
duckduckgo-search==4.4.2
python-whois==0.9.4
